+++ src/Symfony/Component/Validator/Tests/Constraints/DateTimeValidatorTest.php
@@ -36,6 +36,10 @@ class DateTimeValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
+    /**
+     * @group legacy
+     * @expectedDeprecation Validating a \DateTimeInterface with "Symfony\Component\Validator\Constraints\DateTime" is deprecated since version 4.2. Use "Symfony\Component\Validator\Constraints\Type" instead or remove the constraint if the underlying model is already type hinted to \DateTimeInterface.
+     */
     public function testDateTimeClassIsValid()
     {
         $this->validator->validate(new \DateTime(), new DateTime());
@@ -43,6 +47,10 @@ class DateTimeValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
+    /**
+     * @group legacy
+     * @expectedDeprecation Validating a \DateTimeInterface with "Symfony\Component\Validator\Constraints\DateTime" is deprecated since version 4.2. Use "Symfony\Component\Validator\Constraints\Type" instead or remove the constraint if the underlying model is already type hinted to \DateTimeInterface.
+     */
     public function testDateTimeImmutableClassIsValid()
     {
         $this->validator->validate(new \DateTimeImmutable(), new DateTime());